requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
exceptiongroup==1.2.0
greenlet==3.0.1
h11==0.14.0
idna==3.6
lxml==4.9.3
MarkupSafe==2.1.3
MechanicalSoup==1.3.0
outcome==1.3.0.post0
playwright==1.40.0
pyee==11.0.1
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
urllib3==2.1.0
watchdog==3.0.0
Werkzeug==2.0.2
wsproto==1.2.0
